move mock segment server for workers to server dir
